convert to ES6 imports and start to reorganize routes
