+++ server/db.js
@@ -1,16 +1,15 @@
-const mongoose = require("mongoose"); // library to connect to MongoDB
-const logger = require("pino")(); // import pino logger
+import mongoose from "mongoose";
+const { connect } = mongoose;
+import pino from "pino";
+const logger = pino();
 
-module.exports = {
-  init: () => {
-    // connect to mongodb
-    mongoose
-      .connect(process.env.MONGO_CONNECTION_URI, {
-        useNewUrlParser: true,
-        useUnifiedTopology: true,
-        useFindAndModify: false,
-      })
-      .then(() => logger.info("Server connected to MongoDB"))
-      .catch((err) => logger.error("Error connecting to MongoDB", err));
-  },
+export const init = () => {
+  // connect to mongodb
+  connect(process.env.MONGO_CONNECTION_URI, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+    useFindAndModify: false,
+  })
+    .then(() => logger.info("Server connected to MongoDB"))
+    .catch((err) => logger.error("Error connecting to MongoDB", err));
 };
